requirements.txt
covalent-ssh-plugin>=0.17.0,<1
# Workaround https://github.com/ronf/asyncssh/issues/678
asyncssh<2.15
azure-cli
